payment update currency updated

+++ src/components/helper/AlertModalDeposit.jsx
@@ -36,7 +36,7 @@ const AlertModalDeposit = ({
     // setAmount(defaultAmount);
     const defaultAt =
       Number.parseInt(defaultAmount) *
-      Number.parseInt(usercountry.countrycurrencyvaluecomparedtoinr);
+      Number.parseFloat(usercountry.countrycurrencyvaluecomparedtoinr);
     setAmount(defaultAt);
   }, [defaultAmount]);
